repair img create descript ingred & step section
img src on banana breadrecipe page repaired. incorrect file path
and mispelled jpg name of photo. Description of the recipe was added
under the photo.

Unordered list was used to name items needed for the recipe and
ordered lists were used to list the steps taken to make the banana
bread recipe. links to author of main recipe and site used were added
to the bottom of the page.